Move ModsConverter class to Mods.cs, Emoji class  to Embeds.cs
