+++ huisbot/Utilities/Embeds.cs
@@ -581,3 +581,40 @@ internal static class Embeds
     { "mania", new Emoji("mania", 1159888330637463623) },
   };
 }
+
+/// <summary>
+/// Represents a Discord emoji with a name and ID.
+/// </summary>
+public class Emoji
+{
+  /// <summary>
+  /// The name of the emoji.
+  /// </summary>
+  public string Name { get; }
+
+  /// <summary>
+  /// The snowflake ID of the emoji.
+  /// </summary>
+  public ulong Id { get; }
+
+  /// <summary>
+  /// Returns the asset url of this emoji.
+  /// </summary>
+  public string Url => $"https://cdn.discordapp.com/emojis/{Id}.webp";
+
+  /// <summary>
+  /// Returns the emoji string representation of this emoji.
+  /// </summary>
+  public override string ToString() => $"<:{Name}:{Id}>";
+
+  /// <summary>
+  /// Creates a new <see cref="Emoji"/> object with the name and ID of the custom emoji.
+  /// </summary>
+  /// <param name="name">The name of the emoji.</param>
+  /// <param name="id">The ID of the emoji.</param>
+  public Emoji(string name, ulong id)
+  {
+    Name = name;
+    Id = id;
+  }
+}
